Users: Add , Delete, Update, and Get all and get by ID in postman
Photos: Add , Delete, Update, and Get all by id in postman :)

sharedPhotos: Get all and get by id in postman , had to add IDClass

Frontend: add and view, update users working properly

+++ src/main/frontend/src/components/ListUserComponent.jsx
@@ -102,7 +102,12 @@ export default class ListUserComponent extends Component {
                         variant="contained"
                         startIcon={<DeleteIcon />}
                         onClick={() => {
-                          this.deleteUser(user.id);
+                          if (
+                            window.confirm(
+                              "Are you sure you wish to delete this user?"
+                            )
+                          )
+                            this.deleteUser(user.id);
                         }}
                         classname="p-5"
                       >
